requirements.txt
fastapi==0.68.1
uvicorn==0.13.4
sqlalchemy==1.4.11
python-dotenv==0.17.0
pymysql==1.0.2
cryptography==3.4.7
PyJWT==1.7.1
passlib==1.7.2
jinjasql==0.1.8
pandas==1.2.4
ecdsa==0.14.1
pyasn1==0.4.8
python-jose==3.2.0
rsa==4.7.2
bcrypt==3.2.0
sendgrid==6.7.0
python-multipart==0.0.5
multipart
requests